fix cs

+++ bootstrap/pipe.php
@@ -1,4 +1,10 @@
 <?php
+/**
+ * @link https://github.com/ixocreate
+ * @copyright IXOCREATE GmbH
+ * @license MIT License
+ */
+
 declare(strict_types=1);
 
 namespace Ixocreate\Admin;
@@ -35,11 +41,8 @@ use Ixocreate\Application\Http\Pipe\PipeConfigurator;
 use Zend\Expressive\Helper\BodyParams\BodyParamsMiddleware;
 
 /** @var PipeConfigurator $pipe */
-
 $pipe->segmentPipe(AdminConfig::class, 2000000)(function (PipeConfigurator $pipe) {
-
     $pipe->segment('/api')(function (PipeConfigurator $pipe) {
-
         $pipe->pipe(EnforceApiResponseMiddleware::class);
         $pipe->pipe(ErrorMiddleware::class);
         $pipe->pipe(SessionDataMiddleware::class);
@@ -57,44 +60,33 @@ $pipe->segmentPipe(AdminConfig::class, 2000000)(function (PipeConfigurator $pipe
             $group->get('/auth/user', UserAction::class, "admin.api.auth.user");
 
             $group->get('/user', \Ixocreate\Admin\Action\Api\User\IndexAction::class, 'admin.api.user.index');
-            $group->post('/user', \Ixocreate\Admin\Action\Api\User\CreateAction::class,
-                'admin.api.user.create');
-            $group->get('/user/{id}', \Ixocreate\Admin\Action\Api\User\DetailAction::class,
-                'admin.api.user.detail');
-            $group->get('/user/config', \Ixocreate\Admin\Action\Api\User\ConfigAction::class,
-                'admin.api.user.config');
-            $group->patch('/user/{id}', \Ixocreate\Admin\Action\Api\User\UpdateAction::class,
-                'admin.api.user.update');
-            $group->delete('/user/{id}', \Ixocreate\Admin\Action\Api\User\DeleteAction::class,
-                'admin.api.user.delete');
+            $group->post('/user', \Ixocreate\Admin\Action\Api\User\CreateAction::class, 'admin.api.user.create');
+            $group->get('/user/{id}', \Ixocreate\Admin\Action\Api\User\DetailAction::class, 'admin.api.user.detail');
+            $group->get('/user/config', \Ixocreate\Admin\Action\Api\User\ConfigAction::class, 'admin.api.user.config');
+            $group->patch('/user/{id}', \Ixocreate\Admin\Action\Api\User\UpdateAction::class, 'admin.api.user.update');
+            $group->delete('/user/{id}', \Ixocreate\Admin\Action\Api\User\DeleteAction::class, 'admin.api.user.delete');
             $group->post('/user/resetPassword', ChangeEmailAction::class, 'admin.api.user.resetPassword');
 
-            $group->get('/account/config', \Ixocreate\Admin\Action\Account\ConfigAction::class,
-                'admin.api.account.config');
+            $group->get('/account/config', \Ixocreate\Admin\Action\Account\ConfigAction::class, 'admin.api.account.config');
             $group->patch('/account/email', ChangeEmailAction::class, 'admin.api.account.email');
             $group->patch('/account/locale', ChangeLocaleAction::class, 'admin.api.account.locale');
             $group->patch('/account/password', ChangePasswordAction::class, 'admin.api.account.password');
-            $group->patch('/account/attributes', \Ixocreate\Admin\Action\Account\ChangeAttributesAction::class,
-                'admin.api.account.attributes');
+            $group->patch(
+                '/account/attributes',
+                \Ixocreate\Admin\Action\Account\ChangeAttributesAction::class,
+                'admin.api.account.attributes'
+            );
 
-            $group->get('/dashboard', \Ixocreate\Admin\Action\Api\Dashboard\IndexAction::class,
-                'admin.api.dashboard.index');
+            $group->get('/dashboard', \Ixocreate\Admin\Action\Api\Dashboard\IndexAction::class, 'admin.api.dashboard.index');
 
             $group->group('resource')(function (GroupPipeConfigurator $group) {
-                $group->get('/resource/{resource}', \Ixocreate\Admin\Action\Api\Resource\IndexAction::class,
-                    'admin.api.resource.index', PHP_INT_MAX * -1);
-                $group->get('/resource/{resource}/detail/{id}', DetailAction::class, 'admin.api.resource.detail',
-                    PHP_INT_MAX * -1);
-                $group->get('/resource/{resource}/default-values', DefaultValueAction::class,
-                    'admin.api.resource.defaultValue', PHP_INT_MAX * -1);
-                $group->patch('/resource/{resource}/{id}', UpdateAction::class, 'admin.api.resource.update',
-                    PHP_INT_MAX * -1);
-                $group->post('/resource/{resource}', CreateAction::class, 'admin.api.resource.create',
-                    PHP_INT_MAX * -1);
-                $group->delete('/resource/{resource}/{id}', DeleteAction::class, 'admin.api.resource.delete',
-                    PHP_INT_MAX * -1);
-                $group->get('/resource/{resource}/widget/{position}/{type}[/{id}]', WidgetsAction::class,
-                    'admin.api.resource.widgets', PHP_INT_MAX * -1);
+                $group->get('/resource/{resource}', \Ixocreate\Admin\Action\Api\Resource\IndexAction::class, 'admin.api.resource.index', PHP_INT_MAX * -1);
+                $group->get('/resource/{resource}/detail/{id}', DetailAction::class, 'admin.api.resource.detail', PHP_INT_MAX * -1);
+                $group->get('/resource/{resource}/default-values', DefaultValueAction::class, 'admin.api.resource.defaultValue', PHP_INT_MAX * -1);
+                $group->patch('/resource/{resource}/{id}', UpdateAction::class, 'admin.api.resource.update', PHP_INT_MAX * -1);
+                $group->post('/resource/{resource}', CreateAction::class, 'admin.api.resource.create', PHP_INT_MAX * -1);
+                $group->delete('/resource/{resource}/{id}', DeleteAction::class, 'admin.api.resource.delete', PHP_INT_MAX * -1);
+                $group->get('/resource/{resource}/widget/{position}/{type}[/{id}]', WidgetsAction::class, 'admin.api.resource.widgets', PHP_INT_MAX * -1);
             });
         });
     });
